fixes after review comments
